# Internal/dto/update_memper_role.go
package dto

type UpdateMemberRoleDto struct {
	UserID  int64  `json:"user_id" binding:"required"`
	Role    string `json:"role" binding:"required,oneof=Admin Member Viewer"`
	Version int64  `json:"version" binding:"required"`
}