# docs/swagger.yaml
basePath: /
definitions:
  dto.AddMemberDto:
    properties:
      role:
        enum:
        - Admin
        - Member
        - Viewer
        type: string
      user_id:
        type: integer
    required:
    - role
    - user_id
    type: object
...
  dto.CreateWorkspaceRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
...
  dto.UpdateMemberRoleDto:
    properties:
      role:
        enum:
        - Admin
        - Member
        - Viewer
        type: string
      user_id:
        type: integer
      version:
        type: integer
    required:
    - role
    - user_id
    - version
    type: object
...
  dto.UpdateWorkspaceDto:
    properties:
      description:
        type: string
      name:
        type: string
      version:
        type: integer
    required:
    - version
    type: object
...
  model.Workspace:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        format: int64
        type: integer
      name:
        type: string
      ownerID:
        format: int64
        type: integer
      version:
        format: int64
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: mohammed.qassem@baly.iq
    name: Mohammed
  description: Task Management System API
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Task Management API
  version: "1.0"
paths:
...
  /workspace:
    get:
      consumes:
      - application/json
      description: Create a new workSpace
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Workspace'
      security:
      - BearerAuth: []
      summary: CreateWorkspace
      tags:
      - WorkSpace
  /workspace/{id}/Delete:
    delete:
      consumes:
      - application/json
      description: Delete the workSpace
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: DeleteWorkspace
      tags:
      - WorkSpace
  /workspace/{id}/DeleteMember/{user_id}:
    delete:
      consumes:
      - application/json
      description: delete memper from the workspace
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: DeleteMember
      tags:
      - WorkSpaceMemper
  /workspace/{id}/Update:
    put:
      consumes:
      - application/json
      description: update the workSpace
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateWorkspaceDto Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateWorkspaceDto'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: UpdateWorkspace
      tags:
      - WorkSpace
  /workspace/{id}/UpdateMemberRole:
    put:
      consumes:
      - application/json
      description: updete the memper role
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateMemberRoleDto Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMemberRoleDto'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: UpdateMemberRole
      tags:
      - WorkSpaceMemper
  /workspace/{id}/invite:
    post:
      consumes:
      - application/json
      description: Invate memper to the workspace
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      - description: AddMemberDto Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddMemberDto'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: InviteMember
      tags:
      - WorkSpaceMemper
  /workspace/{id}/members:
    get:
      consumes:
      - application/json
      description: Get all the Workspace Members
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: GetWorkspaceMembers
      tags:
      - WorkSpaceMemper
  /workspace/create:
    post:
      consumes:
      - application/json
      description: Create a new workSpace
      parameters:
      - description: CreateWorkspace Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateWorkspaceRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: CreateWorkspace
      tags:
      - WorkSpace
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"